k1om/3.0: MIC interrupt + platform + parallel-boot port (HW-validated boot)
Ports the MIC-specific layer to Linux 3.0 APIs and restores parallel SMP boot.
Validated on real KNC hardware: boots to Gentoo userspace + sshd, 228 CPUs up in
0.46s, EXT4 root on virtblk, SCIF online.

- apic/io_apic.c: SBOX APIC-ICR irq_chip ported to v3.0 irq_data API
  (irq_mask/unmask/ack, irq_set_chip_and_handler_name, irq_set_handler_data,
  irq_get_chip_data), MIC memory-mapped IOAPIC read/write, u64 IOAPIC address,
  MIC_NUM_IOAPIC_ENTRIES; arch_setup_sbox_irqs/setup_sbox_irq (no longer stubbed).
- apic/ipi.c + apic_flat_64.c: physical-mode IPI helpers, MIC 9-bit APIC IDs,
  physflat shorthand.
- smpboot.c: mic_construct_default_ioirq_mptable + mic_smpt_init +
  disable_icache_snoop restored in native_smp_prepare_cpus (were dropped when
  smpboot.c was reset); this repopulates mp_irqs so the RDMASR/DMA IOAPIC irqs
  (incl. SCIF irq 17) get chips.
- Parallel AP boot RESTORED (was reset to serial): do_boot_cpu_parallel /
  native_cpu_up_parallel / do_check_cpu / do_pre_wakeup_aps / kick_secondary_cpu
  (smpboot.c) + cpu_up_parallel / map_cpu_active (kernel/cpu.c) + two-phase
  smp_init (kernel/smp.c). head_64.S per-apicid stack/gs assembly was already
  present. CONFIG_X86_MICPCI now selects VIRTIO/VIRTIO_RING.
- Config: VIRTIO, VIRTIO_RING, EXT4_FS, DEVTMPFS, I2C_CHARDEV, HVC_MIC enabled;
  micpm CORE on (CPUIDLE/CPUFREQ still off); KDB off.

Remaining (see WORKLOG-3.0-port.md): card->host vnet/SCIF delivery is intermittent
(kernel works - initramfs DHCP can succeed - but unreliable; correlates with
MPSSBOOT notify -19). micras (i915 gmbus), micpm cpuidle/cpufreq still deferred.

+++ arch/x86/include/asm/io_apic.h
@@ -169,7 +169,7 @@ extern struct mp_ioapic_gsi  mp_gsi_routing[];
 extern u32 gsi_top;
 int mp_find_ioapic(u32 gsi);
 int mp_find_ioapic_pin(int ioapic, u32 gsi);
-void __init mp_register_ioapic(int id, u32 address, u32 gsi_base);
+void __init mp_register_ioapic(int id, u64 address, u32 gsi_base);
 extern void __init pre_init_apic_IRQ0(void);
 
 extern void mp_save_irq(struct mpc_intsrc *m);
